requirements.txt
absl-py==2.0.0
ansi2html==1.8.0
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
clarabel==0.6.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cvxpy==1.4.1
cycler==0.12.0
dash==2.13.0
dash-antd-components==0.0.1rc2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.5.0
dill==0.3.7
ecos==2.0.12
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.43.0
fsspec==2023.9.2
gast==0.4.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
Grid2Op==1.9.5
grpcio==1.59.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-client==8.3.1
jupyter-core==5.3.2
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
LightSim2Grid==0.7.4
llvmlite==0.41.0
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multiprocess==0.70.15
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.0
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
osqp==0.6.3
outcome==1.2.0
packaging==23.2
pandapower==2.13.1
pandas==2.0.3
parso==0.8.3
pathos==0.2.9
percy==2.0.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
plotly==5.17.0
pluggy==1.3.0
pox==0.3.3
ppft==1.7.6.7
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qdldl==0.1.7.post0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
scs==3.2.3
seaborn==0.11.2
selenium==4.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
triton==2.1.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.6
waitress==2.1.2
wcwidth==0.2.8
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.0.0
zipp==3.17.0
